+++ stroom-query-common/src/main/java/stroom/query/common/v2/ResultHandler.java
@@ -22,16 +22,11 @@ import stroom.util.shared.HasTerminate;
 import java.util.Map;
 
 public interface ResultHandler {
-
     void handle(Map<CoprocessorKey, Payload> payloadMap, HasTerminate hasTerminate);
 
     boolean shouldTerminateSearch();
 
-    boolean isComplete();
-
-    void setComplete(boolean complete);
-
-    void registerCompletionListener(final CompletionListener completionListener);
+    CompletionState getCompletionState();
 
     Data getResultStore(String componentId);
 }