+++ stroom-query-common/src/main/java/stroom/query/common/v2/SearchResultHandler.java
@@ -16,32 +16,19 @@
 
 package stroom.query.common.v2;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import stroom.mapreduce.v2.UnsafePairQueue;
 import stroom.query.api.v2.TableSettings;
 import stroom.query.common.v2.CoprocessorSettingsMap.CoprocessorKey;
-import stroom.query.util.LambdaLogger;
-import stroom.query.util.LambdaLoggerFactory;
 import stroom.util.shared.HasTerminate;
 
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Objects;
-import java.util.Queue;
-import java.util.concurrent.ConcurrentLinkedQueue;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 public class SearchResultHandler implements ResultHandler {
-    private static final Logger LOGGER = LoggerFactory.getLogger(SearchResultHandler.class);
-    private static final LambdaLogger LAMBDA_LOGGER = LambdaLoggerFactory.getLogger(SearchResultHandler.class);
-
     private final CoprocessorSettingsMap coprocessorSettingsMap;
     private final Map<CoprocessorKey, TablePayloadHandler> handlerMap = new HashMap<>();
-    private final AtomicBoolean complete = new AtomicBoolean();
-    private final Queue<CompletionListener> completionListeners = new ConcurrentLinkedQueue<>();
+    private final CompletionState completionState = new CompletionStateImpl();
 
     public SearchResultHandler(final CoprocessorSettingsMap coprocessorSettingsMap,
                                final Sizes defaultMaxResultsSizes,
@@ -111,66 +98,56 @@ public class SearchResultHandler implements ResultHandler {
     }
 
     @Override
-    public boolean isComplete() {
-        return complete.get();
+    public CompletionState getCompletionState() {
+        return completionState;
     }
 
-    private boolean areHandlersBusy() {
-        for (final TablePayloadHandler handler : handlerMap.values()) {
-            if (handler.busy()) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public void setComplete(final boolean complete) {
-        final boolean isAlreadyComplete = this.complete.get();
-        LOGGER.trace("setComplete({}), currentValue={}", complete, isAlreadyComplete);
-        if (isAlreadyComplete && !complete) {
-            throw new RuntimeException("Attempting to mark SearchResultHandler as not complete when it has " +
-                    "already been completed");
-        }
-
-        if (complete) {
-            // We have been told the search is complete but the TablePayloadHandlers may still be doing work
-            // so wait for them
-            boolean hasPendingWorkFinished = false;
-            for (final TablePayloadHandler handler : handlerMap.values()) {
-                hasPendingWorkFinished = handler.waitForPendingWork(10, TimeUnit.SECONDS);
-
-                if (!hasPendingWorkFinished) {
-                    LOGGER.trace("Work still pending after timeout");
-                    break;
-                }
-            }
-            LOGGER.trace("isPendingWorkFinished={}", hasPendingWorkFinished);
-
-            if (hasPendingWorkFinished) {
-                LOGGER.trace("setting complete to {}", true);
-                this.complete.set(true);
-                //notify the listeners
-                for (CompletionListener listener; (listener = completionListeners.poll()) != null; ) {
-                    // when notified they will check isComplete
-                    LOGGER.debug("Notifying {} {} that we are complete", listener.getClass().getName(), listener);
-                    listener.onCompletion();
-                }
-            } else {
-                LOGGER.trace("Handlers are busy so not setting complete");
-            }
-        }
-    }
-
-    @Override
-    public void registerCompletionListener(final CompletionListener completionListener) {
-        if (complete.get()) {
-            //immediate notification
-            completionListener.onCompletion();
-        } else {
-            completionListeners.add(Objects.requireNonNull(completionListener));
-        }
-    }
+//    private boolean areHandlersBusy() {
+//        for (final TablePayloadHandler handler : handlerMap.values()) {
+//            if (handler.busy()) {
+//                return true;
+//            }
+//        }
+//        return false;
+//    }
+//
+//    @Override
+//    public void setComplete(final boolean complete) {
+//        final boolean isAlreadyComplete = this.complete.get();
+//        LOGGER.trace("setComplete({}), currentValue={}", complete, isAlreadyComplete);
+//        if (isAlreadyComplete && !complete) {
+//            throw new RuntimeException("Attempting to mark SearchResultHandler as not complete when it has " +
+//                    "already been completed");
+//        }
+//
+//        if (complete) {
+//            // We have been told the search is complete but the TablePayloadHandlers may still be doing work
+//            // so wait for them
+//            boolean hasPendingWorkFinished = false;
+//            for (final TablePayloadHandler handler : handlerMap.values()) {
+//                hasPendingWorkFinished = handler.waitForPendingWork(10, TimeUnit.SECONDS);
+//
+//                if (!hasPendingWorkFinished) {
+//                    LOGGER.trace("Work still pending after timeout");
+//                    break;
+//                }
+//            }
+//            LOGGER.trace("isPendingWorkFinished={}", hasPendingWorkFinished);
+//
+//            if (hasPendingWorkFinished) {
+//                LOGGER.trace("setting complete to {}", true);
+//                this.complete.set(true);
+//                //notify the listeners
+//                for (CompletionListener listener; (listener = completionListeners.poll()) != null; ) {
+//                    // when notified they will check isComplete
+//                    LOGGER.debug("Notifying {} {} that we are complete", listener.getClass().getName(), listener);
+//                    listener.onCompletion();
+//                }
+//            } else {
+//                LOGGER.trace("Handlers are busy so not setting complete");
+//            }
+//        }
+//    }
 
     @Override
     public Data getResultStore(final String componentId) {